voice-assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
openai-whisper==20231117
openai==1.3.0
anthropic==0.8.0
aiohttp==3.9.0
torch
torchaudio
pydantic==2.5.0
python-multipart==0.0.6